+++ packages/shared/src/index.ts
@@ -2,5 +2,6 @@ export * from './logging/logger';
 export * from './logging/http-request';
 export * from './logging/nest-logger';
 export * from './http/access-log.middleware';
+export * from './http/validation';
 export * from './security/headers';
 export * from './types/domain';